pyproject.toml
[tool.poetry]
name = "sicbackend"
version = "0.1.1"
description = "Restful api backend for Sic"
authors = ["blcklptn"]
license = "Fuck You"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
sqlalchemy = "^2.0.4"
psycopg2-binary = "^2.9.5"
pydantic = "^1.10.5"
bcrypt = "^4.0.1"
pyjwt = "^2.6.0"
uvicorn = "^0.20.0"
python-jose = "^3.3.0"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
